requirements.txt
astroid==1.5.3
certifi==2017.11.5
chardet==3.0.4
clearbit==0.1.7
decorator==4.1.2
Django==1.11.7
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
email-hunter-python==1.1.0
Faker==0.8.7
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.7.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
